requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.9.7
djangorestframework==3.3.3
ipython==5.0.0
ipython-genutils==0.1.0
mysqlclient==1.3.7
pexpect==4.2.0
pickleshare==0.7.2
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
